# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HealthAI Coach API",
    "description": "\n    API REST du backend métier HealthAI Coach.\n\n    Gère les données de nutrition, fitness et utilisateurs\n    pour la plateforme de coaching santé personnalisé.\n    ",
    "contact": {
      "name": "Équipe HealthAI Coach"
    },
    "version": "1.0.0"
  },
  "paths": {
    "/users/": {
      "get": {
        "tags": [
          "Utilisateurs"
        ],
        "summary": "Lister les utilisateurs",
        "description": "Retourne la liste paginée de tous les utilisateurs enregistrés.",
        "operationId": "list_users_users__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Nombre d'entrées à ignorer",
              "default": 0,
              "title": "Skip"
            },
            "description": "Nombre d'entrées à ignorer"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Nombre maximum d'entrées à retourner",
              "default": 100,
              "title": "Limit"
            },
            "description": "Nombre maximum d'entrées à retourner"
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtre de recherche sur le nom/prénom",
              "title": "Q"
            },
            "description": "Filtre de recherche sur le nom/prénom"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponse"
                  },
                  "title": "Response List Users Users  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Utilisateurs"
        ],
        "summary": "Créer un utilisateur",
        "description": "Crée un nouveau profil utilisateur. L'IMC est recalculé automatiquement côté serveur.",
        "operationId": "create_user_users__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{user_id}": {
      "get": {
        "tags": [
          "Utilisateurs"
        ],
        "summary": "Récupérer un utilisateur",
        "description": "Retourne le profil d'un utilisateur par son identifiant.",
        "operationId": "get_user_users__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Utilisateurs"
        ],
        "summary": "Mettre à jour un utilisateur",
        "description": "Met à jour tous les champs d'un utilisateur existant.",
        "operationId": "update_user_users__user_id__put",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Utilisateurs"
        ],
        "summary": "Supprimer un utilisateur",
        "description": "Supprime un utilisateur et toutes ses données liées (CASCADE).",
        "operationId": "delete_user_users__user_id__delete",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/register": {
      "post": {
        "tags": [
          "Authentification"
        ],
        "summary": "Créer un compte",
        "operationId": "register_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Authentification"
        ],
        "summary": "Se connecter — retourne un token JWT",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_login_auth_login_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "Authentification"
        ],
        "summary": "Profil de l'utilisateur connecté",
        "operationId": "me_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "Authentification"
        ],
        "summary": "Modifier son propre profil",
        "operationId": "update_me_auth_me_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Authentification"
        ],
        "summary": "Supprimer définitivement son compte (RGPD)",
        "description": "Suppression totale et définitive du compte de l'utilisateur connecté.\n\nConforme au droit à l'effacement (art. 17 RGPD). Supprimer la ligne `users`\nliée déclenche, via les contraintes ON DELETE CASCADE, l'effacement de\ntoutes les données associées : compte d'authentification (user_auth),\nprofil santé, métriques biométriques, séances d'entraînement (et exercices\nde séance), journaux alimentaires et publications.\nSi aucun profil santé n'est lié, seul le compte user_auth est supprimé.",
        "operationId": "delete_my_account_auth_me_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/auth/users/": {
      "get": {
        "tags": [
          "Authentification"
        ],
        "summary": "Liste tous les comptes (avec profil santé lié)",
        "operationId": "list_users_auth_users__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserAuthPublic"
                  },
                  "title": "Response List Users Auth Users  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/users/{user_auth_id}": {
      "get": {
        "tags": [
          "Authentification"
        ],
        "summary": "Récupérer un compte par son ID",
        "operationId": "get_user_auth_users__user_auth_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_auth_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Auth Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAuthPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/me/profile": {
      "get": {
        "tags": [
          "Authentification"
        ],
        "summary": "Récupérer le profil santé de l'utilisateur connecté",
        "operationId": "get_my_profile_auth_me_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "Authentification"
        ],
        "summary": "Mettre à jour son profil santé (IMC recalculé automatiquement)",
        "operationId": "update_my_profile_auth_me_profile_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UserCreate": {
        "properties": {
          "age": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age",
            "description": "Âge en années (18–100 si renseigné)"
          },
          "weight_kg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Kg",
            "description": "Poids en kg (20–300 si renseigné)"
          },
          "height_cm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height Cm",
            "description": "Taille en cm (90–230 si renseignée)"
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "title": "Name"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "male",
                  "female",
                  "other"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "bmi": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bmi",
            "description": "Ignoré à l'écriture : recalculé côté serveur à partir du poids et de la taille"
          },
          "body_fat_pct": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Body Fat Pct"
          },
          "goal": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "weight_loss",
                  "muscle_gain",
                  "sleep_improvement",
                  "maintenance"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Goal"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "UserCreate",
        "description": "Données requises pour créer ou mettre à jour un utilisateur (POST / PUT /users)."
      },
      "UserResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "title": "Name"
          },
          "age": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "male",
                  "female",
                  "other"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "weight_kg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Kg"
          },
          "height_cm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height Cm"
          },
          "bmi": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bmi"
          },
          "body_fat_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body Fat Pct"
          },
          "goal": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "weight_loss",
                  "muscle_gain",
                  "sleep_improvement",
                  "maintenance"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Goal"
          },
          "created_at": {
            "type": "string",
            "format": "date",
            "title": "Created At"
          },
          "profile_issues": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Profile Issues",
            "description": "Champs biométriques hors limites — à corriger via le formulaire profil"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "created_at"
        ],
        "title": "UserResponse",
        "description": "Données renvoyées par l'API (GET). Accepte les valeurs legacy ; signale les anomalies."
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        }
      }
    }
  }
}